requirements.txt
Django==5.2.11
python-dotenv==1.2.1
mysqlclient==2.2.8
Pillow==12.1.1
django-nested-admin==4.1.1
